Removes legacy GlobalLabel class
Removes the last usage of GlobalLabel. This changes the dashboard labels
endpoint to get a unique list of label titles using SQL.

GlobalLabel used to be how we implemented group labels where we grouped
project labels by title. For this dashboard endpoint, we're not showing
group labels but we're only returning labels that have unique titles
because these are used by the filter bar and we filter by title so
duplicates are redundant.