Make the repository read-only while running cleanup
We want to perform some potentially-racy actions on the repository, so
we need to make it read-only while they run to rule out any chance of
data loss.

Ideally, Gitaly would be able to enforce the read-only status itself,
but we really need mandatory praefect before we can do it there.