Fix CI config viz for cross-stage lines
This apply a new way to trace lines in the
CI config visualization if the line depends
on a dependency across multiple stages.
This will let the line stay straight until it
reaches a point just before the node it connects
to, and then draw the Bezier as it used to.