Raise Puma Worker Killer RAM limits
On a 2-worker node, the previous limits were set to:

550 + (2 * 850) = 2250 MB

However, on an Omnibus 13.5.x install, the baseline memory usage appears
to be 800 MB for both the main and worker processes for a total of 2400
MB. Every request is thus over the limit.

On GitLab.com, `puma_per_worker_max_memory_mb` was set to 1342 MB
because we were observing frequent out of memory kills, so we use the
defaults discussed in
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/8760 and
set in
https://ops.gitlab.net/gitlab-cookbooks/chef-repo/-/merge_requests/2352.
We use a slightly lower number since not all GitLab installations will
have as much memory.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/267499