Fix a race condition checking whether a project is read-only
Two problems here: checks for the same function can race, but also,
checks for two different functions can race with each other.

If we allow a hashed storage migration job to run at the same time as a
shard move job, it's hard to saw what would happen. Best avoided.

This change is being made to rationalise a third user of the read-only
mechanism before it is introduced.