Add batching array resolver
This adds a new kind of resolver that is able to batch requests for
collections of items, resolving to an array of those results.

This includes changes to the presenter instrumentation to accept only
known values from the context, and some minor changes to the
graphql_helpers to help constructing resolvers.