Gracefully recover from deleted LFS file
If an LFS object exists in the database but the underlying file has been
deleted, pushing that same object from another project may result in a
confusing "403 Access Forbidden" error.

Now, if the file no longer exists, we replace the file entry with the
newly-uploaded file so that we can gracefully recover from this case.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/25852