Limit the number of identifiers we create
We are limiting the number of identifiers we create per finding as 20.
This number is not a hard limit which means we are not raising an error
in case if there are findings with more than 20 identifiers.
We are just taking the first 20 identifiers and ignoring the rest if
there are more than that.