Propagate integration to group descendants
After propagate group-level integrations
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/44023, we missed
one scenario where the integration could inherit from another
integration from the ancestors, not the closest one, this is something
that we also fixed finding the closest integration
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/45022.

Example: Having a group, a subgroup, and a project belonging to the
subgroup. If we update the group, it propagates the settings to the
subgroup and project (both having inherit_from_id = group.id (pointing
to the group) but if after that, we update the settings of the subgroup,
it doesn't propagate them to the project. This is because the project's
inherit_from_id is still pointing to the group id from the first
propagation.

This always updates the integrations of any descendants for a group
integration.