Add Group migration imports to usage data
We want to be able to report on how many unique users are using the
BulkImport feature, so that we can determine the adoption rate of this
feature.

Add this data to the usage data model that is reported in the usage
ping.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/249169
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/45536