Remove Use of jQuery Tooltip in Modal Copy Button
This currently relies on a Bootstrap root event to hide the tooltip, as
code in `app/assets/javascripts/behaviors/copy_to_clipboard.js` shows a
new tooltip with the `Copied` text on success.

Until this is changed so the tooltip state is all local, the easiest way
is to use a root event.