Refactor to enable not all records are replicable
Some replicable types might have records that are not possible to
replicate.

This change splits out a scope for the replicables that are
replicatable. The Geo primary node can use this scope for example for
checksumming. To determine replicables for a Secondary node, this scope
is also used, in combination with the selective sync and object storage
settings.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/243567