Set flag well before testing its effect
This test fails intermittently, behaving as if the feature flag isn't
enabled even though the API reports that it is.

The feature flag was set immediately before the page was loaded to
check the effect of the feature flag, so this could be a race
condition in which the API acknowledges the flag as enabled, but Rails
generates the page based on stale settings.

This changes the order of steps so that the feature flag is set as
early as possible (after the project is created), yo allow time for
possible stale flag values to expire.