Include experiment key in subject index calc
Before this change, we always bucketed the same users into the same
groups (control or experimental) for the same percentage rollout
regardless of the experiment key. After this change, the same user may
end up in the experimental group for one experiment and the control
group for a different experiment even at the same percentage rollout.