Fix for Security Reports Spec group dashboard
This is a fix for issue 238848
It was found that the group security
dashboard could be in a state that
did not match the test expectations.

After investigation, it was noted that
the group was a shared group that was
used by previous tests. Therefore the
group was in an unknown state.

This MR ensures that this test uses
a group specific to this test