Fix project callbacks when propagating integrations
After introducing the propagation of integrations using batching and
queues https://gitlab.com/gitlab-org/gitlab/-/merge_requests/42128, we
are not updating the project callbacks: has_external_issue_tracker and
has_external_wiki.

The reason for that is that the batch was uncached but after the merge
request batch is scoped to the query.

In this case, the order is important because we do two things:
- Create the integration for those projects without integration.
- Update has_external_issue_tracker and has_external_wiki in those
  projects.

If we first create the integrations when we then try to update the
callbacks, there is nothing to update since the batch scope is empty at
this time. We have to update first and then create the integrations.