Add Caching to BitBucket Server Import for pull requests
The BitBucket Server Import is enhanced with Caching functionality,
this ensures, that already imported Pull Requests will not be imported
or attempted to be imported again. This is crucial as the
SidekiqMemoryKiller can interrupt the process and it will start from
scratch again.

relates to https://gitlab.com/gitlab-org/gitlab/-/issues/23586