Update project callbacks if integration is active
Originally, we only call here to propagate service templates if they are
active. Now, we are propagating instance-level integrations and
group-level integrations (active or not).

has_external_issue_tracker and has_external_wiki assume that the
integration is active. So, we only need to update the project callbacks
if the integration is active.