Remove '@' from Snippet.full_path
We originally added this out of concern that the `snippets` component
in the path could clash with groups or projects named `snippets`, which
we allow for projects and child groups, and only forbid for toplevel
groups.

But we already use the path without '@' for SSH/HTTP clone URLs, and
also include the snippet ID, so this change should be unlikely to cause
problems in practice.

This also makes snippets more consistent with the other `HasRepository`
classes, and allows us to remove some special handling for snippets.