Fix viewing some legacy diff notes in discussions
LegacyDiffNotes created by the GitHub importer lack details about the
SHAs they referenced on the GitHub side. This exposes a bug in our
diff viewer selection logic, where we inappropriately pick the image
viewer whenever both old blob and new blob are nil.

In practice, the image diff viewer relies on both old and new blob to
do its work, so this just leads to us being unable to render the diff.

Adding this condition allows us to fall back to the basic text diff,
which is usually sufficient.