Improve login page provider specs
The previous specs only tested if the amount of tabs matches with the
amount of tabpanes. However, the mocking was not correct when testing
for LDAP or crowd providers. Therefore the test didn't test if we
actually renders tabs for these providers. This can lead to false
positive tests.

This MR fixes the issue and ensures that we actually render the tabs
that we intend to render for the providers.