Fix issue with importable repo list not refreshing
Prior to this change, the filter function on the GitHub repo import page
would set the filter on the state and clear the repos, which would
result in no repos being shown in the filtering. After setting the
filter, we also need to re-fetch the repos from the backend that match
this filter.

Add a call to fetch repos after setting the filter and make the filter
case insensitive.

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/45783
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/36556