Resolve IP address from RequestContext for audit
Callers of AuditEventService does not always have access to IP
address (e.g. Service classes). In addition, acquiring this piece of
information should not be the responsibility of the callers.

IP address can be retrieved from RequestContext and injected directly on
AuditEventService. This approach helps keeping the interface to
AuditEventService small and be more user-friendly.