Add readonly editor to visualize ci config
As a first iteration on the CI editor, this change adds an simple editor
(editor lite) and loads data of a single .gitlab-ci file.

This change uses an graphql queyr that wraps a REST API to load the
contents of the file. Some client side schema definitions were added
for this.