Set feature category in metrics for ETag caching hits
When a request hits a Rails controller, we set the
X-Gitlab-Feature-Category header to the feature category for that path,
which we then use as a label in our Prometheus metrics.

For ETag caching hits, we weren't setting this, because those skip Rails
entirely (that's the point of them!). This annotates the ETag caching
routes with their feature category to set the same header there.

We don't need this for cache misses because they go to Rails anyway.