Refactor StartupJS for GraphQL
This refactors our StartupJS for GraphQL to be a Apollo Link [0] based
application. Instead of manually filling the cache, we build an
interceptor which "short-circuits" the Apollo Link pipeline in case the
query can be found in StartupJS. In case the query fails, is not cached,
has different variables, is done more than once, we skip it down the
pipeline. Also if all Startup Queries have been done, it self-disables.

We also now batch all StartupJS requests into one GraphQL call.

[0]: https://www.apollographql.com/docs/link/overview/