Use DSL style for feature categories in users API
This isn't possible for custom attributes, as they're a mixin, but
everywhere else can use the DSL style.

This also means a slight tweak to the validation in WithFeatureCategory:
previously we didn't allow duplicates at all, but now we allow
duplicates with the same category. That's because the API supports
different HTTP methods to the same path, but those methods will always
be the same category. Omitting the feature category from the duplicates
would look odd.