Download Knapsack and flaky specs report from latest master artifacts
This dogfood artifacts and download the Knapsack report, flaky specs
report, and Crystalball report from the latest successful master
pipeline triggered by @gitlab-bot (this should be a scheduled pipeline
for which we're sure update-tests-metadata was run).

This also remove the need to upload these reports to S3 since they're
already uploaded as artifacts.

Signed-off-by: Rémy Coutable <remy@rymai.me>