Initialize http_requests_total label values
Add http method/status initilaization to the `http_requests_total`
counter. This is required to track error rates and error ratios.

For example, the ratio of 5xx requests to all requests.

    sum(rate(http_requests_total{status=~"5.."}[5m])
    /
    sum(rate(http_requests_total[5m])

Without the inialization, there would be no values for 5xx on small or
low traffic servers. This would lead to blank graphs.

https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/585

Signed-off-by: Ben Kochie <bjk@gitlab.com>