Create new column in historical_data called recorded_at
- Populate recorded_at from created_at or date because created_at
should accurately tell us when the row was recorded. But created_at
is not `null: false` so we need to fallback to date.
- Set a not null constraint on recorded_at.
- Remove the not null constraint on date since we won't be populating
this column anymore. The column will be kept for now in case we need to
go back and check old values. The column can be removed after a
sufficient amount of time has passed with no issues.
- Change the HistoricalData model to use recorded_at instead of date.