Tweaks: Always set state for valid fields, console.warn
A form can safely be reported as valid, without showing validation
feedback. This commit sets field's state to be true if they are valid
to do achieve this.

Also a check has been added to let a developer know if the
required "name" field on an input field is missing.