Update youtube thumbnail URL's for whats_new
Updating the youtube thumbnail urls for all items in data/whats_new. The
only thing I find odd here is that while the url is suggested on forums
like stackoverflow, it's not documented anywhere on youtube. If this url
changes we might need to integrate with youtube and pull the thumbnail
urls from the youtube api.