Checking pipeline subscription availability
This checks whether the pipeline subscription feature
is available before it is rendered. This allows users who don't
have access to this feature to not be presented with a 404 error.

See https://gitlab.com/gitlab-org/gitlab/-/issues/266938

Additionally, this adds **(PREMIUM)** to the docs' markdown, which
improves the visibility of it being a Premium feature.