Add update_sequence_id to JiraConnect workers
In order to make the JiraConnect workers idempotent
we need to make sure they always sent the same
UpdateSequenceId when run with the same arguments.
So the sequence ID needs to be generated at the time
the job is scheduled and not when running the Job.
This is the first part of a multi step release to add
this argument. The next release will remove the default