Make Rails code reloading configurable in development
For many frontend developers, code reloading is unnecessary and creates
more problems than it solves.

This commit adds a CACHE_CLASSES environment variable that can be
defined to disable code reloading. We do this as an environment variable
because this variable is most conveniently defined in the
`development.rb` and isn't needed in production.

Relates to
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/794