Add CI config lint resolver
Refactors to use string as content type

* Refactors mutation to use STRING_TYPE rather than ApolloUpload, since
the mutation cannot work with a path from the frontend
* Removes YamlProcessorResult type and returns StageType instead
* Adds status field to mutation to mirror implementation
* Makes comprehensive specs

Adds a changelog

Filters out jobless stages

Refactored to use custom types

* Added a ci_config type for stages, groups, jobs, needs
* Adds a new gitlab-ci.yml stub that has some includes
* Updates specs
* Updates docs and schema

Updates mutation to be a resolver instead

* Due to some changes, this was better fitted to be resolved in its own
resolver instead of being a mutation. Specs were changed and slightly
modified to reflect this.
* Adds type specs
* Adds some missing fields (size)
* Updates docs and schema
* Updates the changelog