Fix group destroy not working with Gitaly Cluster
Previously when a namespace was deleted, it would always attempt to
remove the legacy storage path. For example, for namespace `test-group`,
Rails would issue a removal for `/path/to/test-group` just in case there
is data there.

If the path does not exist, Gitaly raises an Invalid Argument GRPC (3)
exception, but Gitaly Cluster raises an Internal error (13).

Since hashed storage is now the default and the legacy path shouldn't be
used at all, we can avoid this error entirely by not attempting to
delete the legacy path.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/276501