Ensure that pages metadata exists when updating it
This fix is a bit paranoid: pages_metadata was added in 12.*
and created via background migration.

Now it should exists for every project. (this is true for .com)

But if users just upgrade to 13.* after a long wait,
then background migration may still be in-progress, so we want
to ensure that metadata exist before we update it with
deployment_id