Handle nullbytes in auth headers
When using git-over-http the GitHttpClientController would try to look
up the user or token read from the Authorization headers.

If one of those headers would containe a base64 encoded null-byte,
this would result in an ArgumentError.

This adds support for that to the middleware by decoding the
authorization headers and validating them beforehand.