Merge Request Discussions API: accept commit_id to create discussions on commit
When creating a new thread on a merge request by using the [MR Discussions
API](https://docs.gitlab.com/ee/api/discussions.html#create-new-merge-request-thread),
the thread is always on the merge request's full diff.

This patch allows passing the "commit_id" parameter to associate the thread
with the desired commit.  Additionally, the "commit_id" field of a note
is exposed if the note is a diff note on a merge request.

Use case: for merge requests with many commits, I like to add my threads
on individual commits, so reviewers can immediately find the commit that
introduced the discussed lines.  This can be done in the web interface by
commenting on the diff of a commit within a merge request.  This commit
allows to do the same thing using the API.