Remove re-use of resolver in mutation
Re-using resolvers anywhere is an anti-pattern (and should be completely
banned with a cop) - they do not compose well at all, and calling them
correctly is practically impossible.

And in any case, mutations do not need their
features (laziness, parallel computation) - instead the re-usable
elements are the same as throughout our application: finders and
services.

This replaces the user of a resolver with a finder, and a plain
find_by_full_path, leaving us with simpler code that is easier to follow
and allocates less garbage.

This work was incidental to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/40088
and has been extracted from there.