Extract and move Secret Detection module to CE
This extracts and moves most of a Secret Detection Vuex module from EE
to CE as part of the effort to [add vulnerability counts to the CE MR
security widget][issue].

It's not a straight-forward move, though. The Secret Detection module
contains actions, mutations and getters specific to EE functionality, so
the approach taken here was to move all _common_ behaviour to CE
version, and make the EE version _extend_ the CE version.

A similar approach was taken for the [SAST module][sast_mr].

[issue]: https://gitlab.com/gitlab-org/gitlab/-/issues/273423
[sast_mr]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/47067