Allow feature flag checks to read default_enabled value from YAML
When `default_enabled: :yaml` is passed to the Feature.enabled? method
we read the value from the YAML definition.

With this change we can use the YAML definition as a SSOT. Changing
the default behavior would simply mean flipping the `default_enabled`
in the YAML file, rather than at every occurrence in the code.