Render wiki like issues in search results
Previously we used to render these exactly the same as code results and
only show a few lines of code. Since the wiki is more like human
readable text then it makes more sense to just render it as text in
paragraph form than lines of code. This also allows highlighting to pick
up single words being matched.

Basically this is done by using the same styling we do on issue search
results. It does come with the downside that we truncate the text so
highlighting matching words later in the wiki won't be visible in the
snippet anymore.