Fix N+1 dependency on project deletion
Currently when a project is deleted, it deletes
each of its associated Ci::Build objects one by one.
For projects that rely heavily on CI, this can take a very long time.

This commit changes the dependent destroy method to use :delete_all
instead of :destroy to grately speed up the process.