Move the Rack::Attack config out of initialiser
This moves the config of Rack::Attack out of the initialiser and into
the Gitlab module.

This way, we can avoid monkey-patching `Rack::Attack` but instead call
the configuration on the class. As a result, we can also use the the
normal `prepend_if_ee` strategy to extend rack attack configuration
for EE.