Remove validation from DAST site profile form
DAST site validation has been rethought and will not be part of the DAST
site profile form like it was originally designed to be. This is the
first step in a series of changes to move the validation to a dedicated
modal accessible from the DAST profiles library.

Here, we're simply removing all of the validation logic form the DAST
site profile form to clean things up.