Clear cached merge_ref_sha on reopen
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/41572, we
started caching `merge_ref_sha` when we clean up the merge request
refs.

It is possible to re-open a closed MR though. If we don't clear
the cached `merge_ref_sha`, `MergeRequest#merge_ref_head` will
always point to the cached one which can be stale.

This fixes that by clearing the cached value on reopen.