Update source installation to use Git from Gitaly
Currently, there's quite a lot of different places which keep track
about which Git version is required by Gitaly in order to have it
operate correctly. This list is hard to keep current, and naturally
those places have diverged in the path.

To fix this, we're moving towards a single source of truth for the
current Git version, which is going to be Gitaly. To implement this,
Gitaly nowadays provides a "make git" target which automatically builds
a supported configuration of Git and installs it into the desired
location. While not yet done, the vision is that this allows us to
quickly iterate in case there are security vulnerabilities reported or
in case there's a bug in Git which impacts operations.

As a first target, this commit converts our source-based installation
instructions to use this target.