Update existing records when recording exp. users
Previously we just exited out of the user recording process if we found
an existing experiment_users record. Now we update that existing record.
This is very useful in the case that the rollout percentage changes and
a previously recorded "control" group user moves into the "experimental"
group.

Also, I removed a lot of the extra association setup from the Experiment
model as it wasn't being used anywhere outside of the model and my
updates to that model no longer make use of those associations.