Add doc page for project-level Issue Analytics
Project-level Issue Analytics was introduced after group-level Issue
Analytics, with documentation that simply pointed to the original
group-level documentation. That caused the documentation to be more
complex than it needed to be, and also caused the documentation left-
nav to behave incorrectly. This introduces a separate page for the
project-level feature to fix these problems.