Add custom parallel rspec runner
A custom parallel rspec runner based on Knapsack runner
which takes in additional option for a file containing
list of test files.

When executing RSpec in CI, the list of tests allocated by Knapsack
will be compared with the test files listed in the file.

Only the test files allocated by Knapsack and listed in the file
would be executed in the CI node.