Release Highlights Linter
Adding a linter/validator for the release highlights ("whats new") .yaml
files that will be added to the repo every month.

This code leverages ActiveModel::Validations and ties the errors back to
where the error occurs in the .yaml file.

I decided to actually run the `validate_all!` method in the spec in
which it's tested. Though I do feel like moving this to it's own linter
plugin could be a better place eventually.

part of:
https://gitlab.com/gitlab-org/growth/engineering/-/issues/5403