Github importer - import pull request merged by
The merged by field is not available on the list of pull requests, which
is used to import the merge request. Therefore, for each merged pull
request we need to do another request to Github to fetch this field.

Once fetched, if the user can be mapped to a GitLab user, it's added as
to the `MergeRequest::Metrics#merged_by`. Otherwise, a note is added to
the merge request with the text: 'Merged by "login"', where "login" is
the Github login of the user who merged the pull request in Github.