Fix CSP errors relating to matomo.js
Piwik was renamed to Matomo, and we introduced a settings fallback that
would use the old Piwik settings if available. The problem was that we
were always defining the Matomo config, even if the old settings were
not defined, so the HAML would always include this JavaScript.

We now only fall back to the Piwik settings only if they are defined.