Move ci lint components to pipeline editor
This change moves the components and queries related to CI Lint to the
pipeline editor directory, so they become part of the editor codebase.

Why?

The CI lint feature will be moved completly to the pipeline editor
and this is an initial technical iteration to move the files to their
destination.