Support alternate document root directory
This will be useful for supporting no-downtime upgrades. Admins
attempting to upgrade GitLab via our no-downtime upgrade procedure have
found that CSS and JavaScript often don't load while the upgrade is in
progress. This is because in a mixed deployment scenario with a load
balancer, this can happen:

1. User accesses node version N+1, which then makes a CSS/JS request on
version N.
2. User accesses node version N, which then makes a CSS/JS requests on
version N+1.

In both scenarios, the user gets a 404 since only one version of the
assets exist on a given server.

To fix this, we provide an alternate path where previous and future
assets can be stored.

Relates to https://gitlab.com/gitlab-org/gitlab-workhorse/-/issues/304