Fix styles not loaded when startup_css is enabled
On the Jira Connect app, custom CSS styles were not
loading. This was due to code related to startup_css.
We are using `add_page_specific_style` to include the CSS
which then calls `stylesheet_link_tag_defer`. When
`startup_css` is enabled, the stylesheet is included as
`media: print` which only applies the styles when page is
printed instead of applying them to "all" or "screen".

This change adds a param to the helper which allows us
to disable the `defer` behaviour when not needed and
use that in the jira_connect app.

I also update the order of includes in the jira_connect
layout to group stylesheets and javascripts properly.