Correctly handle indexing issues when transferring group
Previously transferring groups used custom code to trigger indexing
after the projects were updated. But this did not account for checking
visibility_level changes. As such we need to pass through the changed
attributes to maintain_elasticsearch_update when calling it directly so
we can determine if we need to also index the associated issues again.