Add explicity primary key definition to AuditEvent
As a part of the change for partitioning audit_events, the AuditEvent
model will need to have the primary_key explicitly set to use only the
id column. While the database table also includes the partition key
(created_at) in the primary key definition, due to Rails support for
multi-column PK it's easier to simply use the id column, which should
still be unique as it's backed by a sequence.