Update feature_category of workers to code_review
There are some workers that are still tagged under
`source_code_management`.

They are more related to `code_review` so this updates these
workers to be tagged under `code_review`.

This also includes some controllers and APIs that are tagged
under `code_review` but are now under `source_code_management`