Avoid branch name checking when creating a new snippet
There a time window between the moment where we start the
snippet creation commit and when the git hooks are called.

In those git hooks we check that the user can only create
the default branch in the repository. Nevertheless, in that
window time, the default branch might change, resulting
in an invalid git operation.