ci: Only keep relevant files after installing tests dependencies
The _build dir is 211MB (uncompressed) but the executables are moved to
the top gitaly folder after they're compiled, so there's no need to save
the _build dir in the artifacts of the setup-test-env job.

That should speed up the artifacts upload/download phases for
setup-test-env and all the jobs that depend on it.

Signed-off-by: Rémy Coutable <remy@rymai.me>