Extract some project logic into new RepositoryStorageMovable concern
We're going to allow snippets to move the repository storage of
their repositories (and later with group wikiws). For this, we
need to reuse some of the existent functionality in projects.

Therefore, the best way to do it is to extract it into a concern
and include it in snippets.