Speed up pull mirroring by getting updated tags straight from Gitaly
Currently, pull mirroring will fetch the full list of tags for a
repository twice - before and after the pull - and use the diff between
them to work out which new tags have been added.

Instead, we can parse the `git fetch` output to learn when tags have
been added, and use that to avoid pulling the list of tags twice if no
tags have been changed by the fetch.

Ideally, we'd use the output of `git fetch` to determine exactly which
tags had been changed, but that adds significant complexity since the
pointed-to SHAs are not included in the output, and we need that for
the push events we create, etc. Still, optimising the common case seems
like a decent win to me.