Update spec and fixtures
This updates the spec and fixtures to more accurately
reflect what this component is doing.

I have removed most of the fixtures as they were
effectively just asserrting that dompurify "works"

I have added in tests that pertain to the bugs this
mr is addressing. Specifically that SVGs were not
rendering.