Add negatable predicate methods
Prior to this change, the matchers would need
in some cases to wait for the condition to return true and
the negate it. This can add up to 10 seconds each time.

This MR, instead of executing the page object's matcher method, uses
a custom matcher for match_when_negated that reads more naturally
and avoids using negative predicate methods such as has_no_file
in the end to end test.