Check if build trace is corrupted by comparing size
This commit adds support for incrementing a Prometheus counter in case
of detecting a malformed / corrupted build log. We store a bytesize of a
build log in the ci_build_pending_states table and compare it to the
size calculated using stored build trace chunks.

If CRC32 checksum of a build log is invalid and a bytesize differs too,
we increment `corrupted` build logs counter and send an appropritate log
to Sentry.