Support repository moved message with all container types
Rename the `ProjectMoved` push check to `ContainerMoved`, and use the
repository identifier to store messages in Redis instead of the project
ID.

Previously this only worked in a project context, but now supports all
repository containers and also adds the container type in the push
message.

To handle the transition from the previous message key to the new one,
we query and delete both keys in Redis.

Changelog: changed