Fix security report widget for forks
Merge requests from forks run the pipeline in the context of the source
project (i.e., the fork), rather than the target project. This change
makes sure that the correct project is used to get the pipeline jobs
information.

This change also works for same-project merge requests, since in that
case the source project _is_ the target project.