Timely sanitization of initial Epic data
Epic, even though is using the issuable components,
doesn't follow the same pattern of sanitization of the
HTML data at the moment the data is read. This leads to
double-rendering of the Vue applications when description
content fetched with polling doesn't match the initial
one. Even though the difference is purely cosmetic
(stripped out line breaks in the initial description),
Vue still updates state and rerenders the view seeing
that the description is different.

This MR fixes this and unifies the way we sanitize output
for Issues and Epics