WIP: Add current trial status to namespace sidebar
An experiment which adds a current trial status widget to the sidebar
for owners of namespaces which are currently being trialed when those
owners are not also members on any other namespace which currently have
a paid (non-trialed) plan. This new sidebar widget is clickable and
takes the user to the billing page.