Don't run Vale on SVG replacement codes
The SVG replacement codes can't be guaranteed to be
real words, but it doesn't matter. They shouldn't be
linted by Vale. This MR proposes a TokenIgnores
line that skips items matching a regex of **{this}**.

Note that tier badges, blessedly, use **(this)**
instead of curly braces, so we can lint them
separately.