Fix Rails/SaveBang violations in migration specs
Our table helper used in migration specs returns real ActiveRecord
classes, so we can just switch to the bang variants.

Auto-corrected with the following command:

```
rubocop -A --only Rails/SaveBang \
  spec/migrations/**/*.rb \
  ee/spec/migrations/**/*.rb \
  spec/lib/gitlab/background_migration/**/*.rb \
  ee/spec/lib/gitlab/background_migration/**/*.rb
```

In some specs we're not dealing with table helpers, in those cases
an exclusion comment is added back next to the exception.