Fail import state whenever repository import fails
If you mistakenly omit the `.git` extension from an import URL, Gitaly
will return an error message, "The requested URL returned error: 301",
because we disabled following redirections in a security release.

In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/35344, we
disabled retries outright in RepositoryImportWorker, but the import
state would only be failed if the project were an import/export. As a
result, an import by URL would be stalled with an "Import in progress"
forever if a user entered in the wrong URL.

To fix this, we now always fail the import state whenever the repository
failed to clone.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/238214