Add MergeRequest.reviewers GraphQL field
This adds the field `MergeRequest.reviewers`.

Preloads are added to account for the dependent relationships.

Our N+1 testing is improved in a bunch of ways, but removing
confounding factors that are outside the scope of the GraphQL schema,
such as authentication, license loading, and Database read-only checks.

More effort is taken here to correctly isolate the tests from each
other, ensuring we have clean request stores, SQL caches, and
batchloader contexts between runs. We ensure that we are in fact using
the SQL cache.