Migrate button toggle for shared project runners to Vue
To remove ambiguity with the current UI for enabling/disabling shared
runners in project settings, this MR replaces the button with a toggle
element. This is also consistent with our usage of toggles in the other
runner settings.

To do this, the UI has been ported to Vue while the POST endpoint
`toggle_shared_runners` has been modified to return the response in
JSON format.

The changes are gated by the `:vueify_shared_runners_toggle` feature
flag.