Sort commit/compare diff files directory first
When viewing merge request changes while ignoring whitespace
changes or viewing changes for a specific commit in a merge
request, they need to be sorted by directory first.

This is to make diffs and each file browser entries consistent
to prevent confusion.

This uses the same logic we use for sorting merge request diff
files so that was extracted into its own class.