Use Gitaly to fetch overview branches
The problem is that we're fetching all branches and then take 5
the most active and then iterate over the whole array until we get
5 branches with committer-date older than 3 months ago.
If we display the stalest branches on the overview, we'll be able
to fetch 5 the most recent and then fetch 5 the oldest.

We perform 2 requests, but they are smaller and don't depend
on the number of branches.